MULE-12334: Deserialisation breaks for inner class
_ Reverting the refactoring on DefaultComponentLocation
_ Fixing the bug
_ Using org.mule.test.integration.exceptions.ExceptionHandlingTestCase#testUntilSuccessfulInTransactionalScope to test that the fix works
