+++ modules/artifact/src/test/java/org/mule/runtime/module/artifact/serializer/CustomJavaSerializationProtocolTestCase.java
@@ -17,8 +17,8 @@ import static org.mule.runtime.api.message.Message.of;
 import static org.mule.runtime.core.api.lifecycle.LifecycleUtils.initialiseIfNeeded;
 import static org.mule.runtime.module.artifact.classloader.ParentFirstLookupStrategy.PARENT_FIRST;
 import org.mule.runtime.core.api.Event;
-import org.mule.runtime.core.internal.message.InternalMessage;
 import org.mule.runtime.core.api.serialization.SerializationException;
+import org.mule.runtime.core.internal.message.InternalMessage;
 import org.mule.runtime.core.internal.serialization.AbstractSerializerProtocolContractTestCase;
 import org.mule.runtime.module.artifact.classloader.ClassLoaderLookupPolicy;
 import org.mule.runtime.module.artifact.classloader.ClassLoaderRepository;
